pyproject.toml
[tool.poetry]
name = "receipt-recognizer"
version = "0.1.0"
description = ""
authors = ["LewayHackintosh <michael.wei@foxmail.com>"]
readme = "README.md"
packages = [{include = "receipt_recognizer"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.24.0"
azure-ai-formrecognizer = "3.2.0"
pymysql = "^1.1.0"


[tool.poetry.group.dev.dependencies]
watchdog = "^3.0.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
